pyproject.toml
[project]
name = "human-use"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.51.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.8.0",
    "openai>=1.78.0",
    "python-dotenv>=1.1.0",
    "rapidata>=2.23.0",
    "streamlit>=1.45.0",
    "streamlit-oauth>=0.1.11",
]
